libgcc 5.2: Workaround upstream build failure with rpm format
If the rpm package format is selected the gcc v5.2 libgcc recipe
will fail in the do_package_write_rpm() task. This is issue #2431
in the Linaro bugzilla [1]. It is caused by the '-2' that is
appended to the end of the release string.

Until this is fixed upstream or gcc v5.3 is adopted in a later
Yocto Project carry a workaround.

[1] https://bugs.linaro.org/show_bug.cgi?id=2431